Refactor RootViewModel
